add Util class(check_id)
